+++ inst/bs4Dash-0.2.0/bs4Dash.js
@@ -9,19 +9,7 @@ $(function () {
     zIndex              : 999999
   });
   $('.connectedSortable .card-header, .connectedSortable .nav-tabs-custom').css('cursor', 'move');
-
-  //// function to get the index of an element in a list
-  //function getIndex(elem) {
-  //  var $t = $(elem);
-  //  return $t.parent().index();
-  //}
-  // select all nav items links that have not the has-treeview class
-  //var $tabs = $( " #mymenu li.nav-item:not(.has-treeview) a" );
-  var tabs = $("#mymenu a[data-toggle='tab']");
-  var selectedTab = tabs.filter(".active.show");
-  //var $selectedTabLink = $selectedTab.attr("id");
-  //var $index = getIndex($selectedTab);
-  //var $indexPane = $index - 1;
+ 
   
   // handle shinyapps.io: w need to extract the worker id and
   // paste it in the url so that the apps works correctly
@@ -35,12 +23,4 @@ $(function () {
     window.history.replaceState( {} , 'newpath', newpath);
   }
   
-  if (selectedTab.length === 0) {
-    // If no tab starts selected, use the first one, if present
-    $(tabs[0]).tab('show');
-    $('.container-fluid.tab-pane:eq(0)').addClass('active show');
-  } //else {
-    //$("#mymenu .nav-item:eq(" + $indexPane + ") a").tab("show");
-    //$(".tab-pane:eq(" + $indexPane + ")").addClass('active show');
-  //}
 });