+++ inst/bs4Dash-0.2.0/leftSidebar.js
@@ -1,5 +1,48 @@
 $(function () {
   
+  // This code makes sure that each time
+    // a tabItem is clicked, we set the value
+    // of the input binding
+var setInputBindingValue = function() {
+  // Trigger event for the tabItemInputBinding
+  var $obj = $('.sidebarMenuSelectedTabItem');
+  var inputBinding = $obj.data('shiny-input-binding');
+  if (typeof inputBinding !== 'undefined') {
+    inputBinding.setValue($obj, $(this).attr('data-value'));
+    $obj.trigger('change');
+  }
+};
+
+$(document).on('shown.bs.tab', '#mymenu a[data-toggle="tab"]',
+               setInputBindingValue);
+  
+  
+  // When document is ready, if there is a sidebar menu with no activated tabs,
+// activate the one specified by `data-start-selected`, or if that's not
+// present, the first one.
+var ensureActivatedTab = function() {
+    // get the selected tabs
+  var tabs = $("#mymenu a[data-toggle='tab']");
+  var selectedTab = tabs.filter(".active.show");
+  
+  if (selectedTab.length === 0) {
+    // If no tab starts selected, use the first one, if present
+    $(tabs[0]).tab('show');
+    $('.container-fluid.tab-pane:eq(0)').addClass('active show');
+    // This is indirectly setting the value of the Shiny input by setting
+    // an attribute on the html element it is bound to. We cannot use the
+    // inputBinding's setValue() method here because this is called too
+    // early (before Shiny has fully initialized)
+    $('.sidebarMenuSelectedTabItem').attr('data-value',
+      $(tabs[0]).attr('data-value'));
+  
+  
+  
+  } 
+};
+
+ensureActivatedTab();
+  
   // select the first sidebar item by default
   var firsSidebarItem = $("#mymenu").children[0];
   if (firsSidebarItem != "undefined") {
@@ -72,4 +115,58 @@ $(function () {
     var treeViewLiLink = treeViewLi.children[0];
     $(treeViewLiLink).addClass("active");
   });
+  
+  
+    //---------------------------------------------------------------------
+// Source file: ../srcjs/input_binding_tabItem.js
+
+/* global Shiny */
+
+// tabItemInputBinding
+// ------------------------------------------------------------------
+// Based on Shiny.tabItemInputBinding, but customized for tabItems in
+// shinydashboard, which have a slightly different structure.
+var tabItemInputBinding = new Shiny.InputBinding();
+$.extend(tabItemInputBinding, {
+  find: function(scope) {
+    return $(scope).find('.sidebarMenuSelectedTabItem');
+  },
+  getValue: function(el) {
+    var value = $(el).attr('data-value');
+    if (value === "null") return null;
+    return value;
+  },
+  setValue: function(el, value) {
+    var self = this;
+    var anchors = $(el).parent('#mymenu').find('li:not(.treeview)').children('a');
+    anchors.each(function() { // eslint-disable-line consistent-return
+      if (self._getTabName($(this)) === value) {
+        //$(this).tab('show');
+        $(el).attr('data-value', self._getTabName($(this)));
+        return false;
+      }
+    });
+  },
+  receiveMessage: function(el, data) {
+    if (data.hasOwnProperty('value'))
+      this.setValue(el, data.value);
+  },
+  subscribe: function(el, callback) {
+    // This event is triggered by deactivateOtherTabs, which is triggered by
+    // shown. The deactivation of other tabs must occur before Shiny gets the
+    // input value.
+    $(el).on('change.tabItemInputBinding', function() {
+      callback();
+    });
+  },
+  unsubscribe: function(el) {
+    $(el).off('.tabItemInputBinding');
+  },
+  _getTabName: function(anchor) {
+    return anchor.attr('data-value');
+  }
+});
+
+Shiny.inputBindings.register(tabItemInputBinding);
+  
 });